pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-promptguard"
version = "0.1.0"
description = "PromptGuard: Enterprise AI Security Sandbox for evaluating LLM robustness against prompt injection attacks"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.111.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "promptguard.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["promptguard", "promptguard.server"]
package-dir = { "promptguard" = ".", "promptguard.server" = "server" }
